requirements.txt
# CrewAI Core
crewai>=0.28.0
crewai-tools>=0.2.0

# IFC Processing
ifcopenshell>=0.7.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization (using matplotlib and PIL as FreeCAD alternatives)
matplotlib>=3.7.0
Pillow>=10.0.0

# Reporting
jinja2>=3.1.0
weasyprint>=60.0

# CLI
click>=8.1.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
